# Optimize type selector by using pattern match